remove code

+++ src/query/service/src/api/rpc/exchange/exchange_transform_shuffle.rs
@@ -89,7 +89,7 @@ struct OutputsBuffer {
 
 impl OutputsBuffer {
     pub fn create(capacity: usize, outputs: usize) -> OutputsBuffer {
-        let capacity = capacity * 1000 as usize;
+        let capacity = capacity * 1000;
         OutputsBuffer {
             inner: vec![capacity; outputs]
                 .into_iter()
